[feat] wish 입력 검증 추가 및 알림 수신 설정 제거
  - wish name 15자 제한, amount 10자리 제한 추가 (amount 타입 Long 변경)
  - wish 응답에서 불필요한 createdAt 필드 제거
  - 알림 수신 여부(notificationEnabled) 백엔드 관리 제거 (앱에서 관리)
  - notificationMinute 0/30 검증을 Member 엔티티로 이동
  - NotificationTimeRequest 검증 메시지 국제화 적용

+++ src/main/java/com/example/nevera/entity/WishEntity.java
@@ -25,7 +25,7 @@ public class WishEntity {
     private String name;
 
     @Column(nullable = false)
-    private int amount;
+    private Long amount;
 
     @Column(name = "created_at", updatable = false, nullable = false)
     private OffsetDateTime createdAt;
@@ -35,7 +35,7 @@ public class WishEntity {
         this.createdAt = OffsetDateTime.now();
     }
 
-    public void update(String name, int amount) {
+    public void update(String name, Long amount) {
         this.name = name;
         this.amount = amount;
     }
